fix(error): Fix all of the compilation error from conversion to in to size_t

+++ common/ECS/entity/registry/ARegistry.tpp
@@ -20,7 +20,7 @@ void ARegistry::registerComponent()
 }
 
 template <typename T>
-void ARegistry::addComponent(int entityId, std::shared_ptr<T> component)
+void ARegistry::addComponent(size_t entityId, std::shared_ptr<T> component)
 {
     const char *typeName = typeid(T).name();
     auto it = _components.find(typeName);
@@ -31,7 +31,7 @@ void ARegistry::addComponent(int entityId, std::shared_ptr<T> component)
 }
 
 template <typename T>
-std::shared_ptr<T> ARegistry::getComponent(int entityId) const
+std::shared_ptr<T> ARegistry::getComponent(size_t entityId) const
 {
     const char *typeName = typeid(T).name();
     auto it = _components.find(typeName);
@@ -43,7 +43,7 @@ std::shared_ptr<T> ARegistry::getComponent(int entityId) const
 }
 
 template <typename T>
-void ARegistry::removeComponent(int entityId)
+void ARegistry::removeComponent(size_t entityId)
 {
     const char *typeName = typeid(T).name();
     auto it = _components.find(typeName);
@@ -54,7 +54,7 @@ void ARegistry::removeComponent(int entityId)
 }
 
 template <typename T>
-bool ARegistry::hasComponent(int entityId) const
+bool ARegistry::hasComponent(size_t entityId) const
 {
     const char *typeName = typeid(T).name();
     auto it = _components.find(typeName);
